+++ t/lib/FSM/Test/RegressionCorpus.pm
@@ -1,0 +1,71 @@
+package FSM::Test::RegressionCorpus;
+
+use strict;
+use warnings;
+
+use Exporter 'import';
+
+our @EXPORT_OK = qw(regression_corpus_entries protocol_fixture_entries);
+
+my @REGRESSION_CORPUS = (
+    {
+        id => 'protocol.apb_requester',
+        relpath => 'fsm/apb_requester.fsm',
+        family => 'protocol_fixture',
+        classification => 'supported_smoke',
+        coverage => 'direct_root_pipeline_cli',
+        source_kind => 'fsm',
+        expected_module_name => 'apb_requester',
+    },
+    {
+        id => 'protocol.apb_completer',
+        relpath => 'fsm/apb_completer.fsm',
+        family => 'protocol_fixture',
+        classification => 'supported_smoke',
+        coverage => 'direct_root_pipeline_cli',
+        source_kind => 'fsm',
+        expected_module_name => 'apb_completer',
+    },
+    {
+        id => 'protocol.amba_requester',
+        relpath => 'fsm/amba_requester.fsm',
+        family => 'protocol_fixture',
+        classification => 'supported_smoke',
+        coverage => 'direct_root_pipeline_cli',
+        source_kind => 'fsm',
+        expected_module_name => 'amba_requester',
+    },
+    {
+        id => 'protocol.apb_tb',
+        relpath => 'fsm/apb_tb.fsm',
+        family => 'protocol_fixture',
+        classification => 'supported_smoke',
+        coverage => 'composition_top_pipeline_cli',
+        source_kind => 'composition',
+        expected_top_name => 'apb_tb',
+        expected_lane => 'C4',
+        expected_instance_count => 2,
+        expected_child_modules => ['apb_requester', 'apb_completer'],
+    },
+);
+
+sub regression_corpus_entries {
+    return map { _copy_entry($_) } @REGRESSION_CORPUS;
+}
+
+sub protocol_fixture_entries {
+    return map { _copy_entry($_) } grep { $_->{family} eq 'protocol_fixture' } @REGRESSION_CORPUS;
+}
+
+sub _copy_entry {
+    my ($entry) = @_;
+
+    my %copy = %{$entry};
+    if (ref $copy{expected_child_modules} eq 'ARRAY') {
+        $copy{expected_child_modules} = [@{$copy{expected_child_modules}}];
+    }
+
+    return \%copy;
+}
+
+1;